Blockchain storage as interface
