+++ blockchain/storage/leveldb/leveldb_test.go
@@ -1,0 +1,33 @@
+package leveldb
+
+import (
+	"io/ioutil"
+	"os"
+	"testing"
+
+	"github.com/umbracle/minimal/blockchain/storage"
+)
+
+func newStorage(t *testing.T) (*Storage, func()) {
+	path, err := ioutil.TempDir("/tmp", "minimal_storage")
+	if err != nil {
+		t.Fatal(err)
+	}
+	s, err := NewLevelDBStorage(path, nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+	close := func() {
+		if err := os.RemoveAll(path); err != nil {
+			t.Fatal(err)
+		}
+	}
+	return s, close
+}
+
+func TestStorage(t *testing.T) {
+	s, close := newStorage(t)
+	defer close()
+
+	storage.TestStorage(t, s)
+}
