+++ blockchain/testing.go
@@ -9,9 +9,10 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/umbracle/minimal/blockchain/storage"
 	"github.com/umbracle/minimal/chain"
 	"github.com/umbracle/minimal/state"
+
+	"github.com/umbracle/minimal/blockchain/storage/leveldb"
 )
 
 type fakeConsensus struct {
@@ -155,7 +156,7 @@ func NewTestBlockchainWithBlocks(t *testing.T, blocks []*types.Block, receipts [
 
 // NewTestBlockchain creates a new dummy blockchain for testing
 func NewTestBlockchain(t *testing.T, headers []*types.Header) *Blockchain {
-	s, err := storage.NewMemoryStorage(nil)
+	s, err := leveldb.NewMemoryStorage(nil)
 	if err != nil {
 		t.Fatal(err)
 	}
